requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.11.1
altgraph==0.17.3
ansicon==1.89.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asteroid-filterbanks==0.4.0
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
audioread==3.0.0
auto-py-to-exe==2.35.0
av==10.0.0
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.12.2
bleach==6.0.0
blessed==1.20.0
bottle==0.12.25
bottle-websocket==0.2.9
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.4
cmaes==0.9.1
cmake==3.26.4
colorama==0.4.6
colorlog==6.7.0
comm==0.1.3
contourpy==1.0.7
coverage==5.5
croniter==1.4.1
cycler==0.11.0
dateutils==0.6.12
debugpy==1.6.7
decorator==4.4.2
deepdiff==6.3.1
defusedxml==0.7.1
deprecation==2.1.0
dlib==19.24.2
docopt==0.6.2
EasyProcess==1.1
Eel==0.16.0
einops==0.6.1
entrypoint2==1.1
et-xmlfile==1.1.0
exceptiongroup==1.1.2
executing==1.2.0
eyed3==0.9.7
fastapi==0.100.0
fastjsonschema==2.17.1
filelock==3.12.0
filetype==1.2.0
fonttools==4.39.4
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.6.0
future==0.18.3
gevent==22.10.2
gevent-websocket==0.10.1
greenlet==2.0.2
h11==0.14.0
halo==0.0.31
hance==1.0.6
hmmlearn==0.3.0
huggingface-hub==0.15.1
HyperPyYAML==1.2.1
idna==3.4
imageio==2.28.1
imageio-ffmpeg==0.4.8
imbalanced-learn==0.10.1
imblearn==0.0
inquirer==3.1.3
ipykernel==6.24.0
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==8.0.7
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jinxed==1.2.0
joblib==1.3.1
jsonpointer==2.4
jsonschema==4.18.0
jsonschema-specifications==2023.6.1
julius==0.2.7
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
kiwisolver==1.4.4
lazy_loader==0.2
librosa==0.10.0.post2
lightning==2.0.5
lightning-cloud==0.5.37
lightning-utilities==0.9.0
llvmlite==0.40.0
log-symbols==0.0.14
Mako==1.2.4
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.1
more-itertools==9.1.0
MouseInfo==0.1.3
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.5
mss==9.0.1
multidict==6.0.4
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.6.0
nbformat==5.9.1
nest-asyncio==1.5.6
networkx==3.1
noisereduce==2.0.1
notebook==6.5.4
notebook_shim==0.2.3
numba==0.57.0
numpy==1.24.3
omegaconf==2.3.0
openai-whisper @ git+https://github.com/openai/whisper.git@b91c907694f96a3fb9da03d4bbdc83fbcd3a40a4
opencv-python==4.7.0.72
openpyxl==3.1.2
optuna==3.2.0
ordered-set==4.1.0
overrides==7.3.1
packaging==23.1
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
pefile==2023.2.7
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.8.1
pooch==1.6.0
primePy==1.3
proglog==0.1.10
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.23.4
psutil==5.9.5
pure-eval==0.2.2
pyannote.audio @ git+https://github.com/pyannote/pyannote-audio@35be745d3d77058fca1eaf318a3a060a0841bcfe
pyannote.core==5.0.0
pyannote.database==5.0.1
pyannote.metrics==3.2.1
pyannote.pipeline==2.3
PyAudio==0.2.13
pyAudioAnalysis==0.3.14
PyAutoGUI==0.9.54
pycparser==2.21
pydantic==1.10.11
pydub==0.25.1
pygame==2.5.0
PyGetWindow==0.0.9
Pygments==2.15.1
pyinstaller==5.12.0
pyinstaller-hooks-contrib==2023.3
PyJWT==2.7.0
PyMsgBox==1.0.9
pyparsing==3.0.9
pyperclip==1.8.2
PyQt5==5.15.9
pyqt5-plugins==5.15.9.2.3
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
pyqt5-tools==5.15.9.3.3
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
python-dateutil==2.8.2
python-dotenv==1.0.0
python-editor==1.0.4
python-json-logger==2.0.7
python-multipart==0.0.6
pytorch-lightning==2.0.5
pytorch-metric-learning==2.2.0
pytweening==1.0.7
pytz==2023.3
PyWavelets==1.4.1
pywin32==306
pywin32-ctypes==0.2.0
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.1.0
qt5-applications==5.15.2.2.3
qt5-tools==5.15.2.1.3
qtconsole==5.4.3
QtPy==2.3.1
readchar==4.0.5
referencing==0.29.1
regex==2023.6.3
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.4.2
rpds-py==0.8.10
ruamel.yaml==0.17.28
ruamel.yaml.clib==0.2.7
safetensors==0.3.1
scikit-image==0.20.0
scikit-learn==1.3.0
scipy==1.10.1
semver==3.0.1
Send2Trash==1.8.2
sentencepiece==0.1.99
shellingham==1.5.0.post1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.4.1
soxr==0.3.5
speechbrain==0.5.14
spinners==0.0.24
SQLAlchemy==2.0.18
stack-data==0.6.2
starlette==0.27.0
starsessions==1.3.0
sympy==1.12
tabulate==0.9.0
tensorboardX==2.6.1
termcolor==2.3.0
terminado==0.17.1
threadpoolctl==3.1.0
tifffile==2023.4.12
tiktoken==0.3.3
tinycss2==1.2.1
tkVideo==0.1
tkvideoplayer==2.3
tokenizers==0.13.3
toml==0.10.2
torch==2.0.1
torch-audiomentations==0.11.0
torch-pitch-shift==1.2.4
torchaudio==2.0.2
torchmetrics==1.0.0
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.30.2
typer==0.9.0
typing_extensions==4.5.0
tzdata==2023.3
uri-template==1.3.0
urllib3==2.0.3
uvicorn==0.22.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
websockets==11.0.3
whichcraft==0.6.1
widgetsnbextension==4.0.8
yarl==1.9.2
zope.event==4.6
zope.interface==6.0
